+++ packages/client-react/src/App.tsx
@@ -4,11 +4,10 @@ import { ToastProvider } from './store/ToastContext'
 import { useEffect, useState } from 'react'
 import { useApp } from './store/AppContext'
 import { apiService } from './services/ApiService'
-import { useGameStore } from './store/gameStore'
 import HomePage from './pages/HomePage'
-import GamePage from './pages/GamePage'
-import AuthPage from './pages/AuthPage'
-import LobbyPage from './pages/LobbyPage'
+import GamePage from './features/game/pages/GamePage'
+import AuthPage from './features/auth/pages/AuthPage'
+import LobbyPage from './features/lobby/pages/LobbyPage'
 import { ErrorBoundary } from './components/ErrorBoundary'
 import './App.css'
