+++ packages/client-react/src/__tests__/components/board-types.test.ts
@@ -4,7 +4,7 @@ import {
   createSampleBoard,
   gridToWorldCoords,
   worldToGridCoords,
-} from '../../components/game/board-types'
+} from '../../features/game/components/board-types'
 
 describe('createEmptyBoard', () => {
   it('should create a 5x5 board by default', () => {
