+++ packages/client-react/src/__tests__/types/gameLifecycle.test.ts
@@ -3,7 +3,7 @@ import {
   mapToGameLifecycle,
   getGameStatusMessage,
   getUIComponents,
-} from '../../types/gameLifecycle'
+} from '../../features/game/types/gameLifecycle'
 
 describe('mapToGameLifecycle', () => {
   describe('FINISHED state', () => {
