+++ packages/client-react/src/features/debug/components/DebugAxis.tsx
