+++ packages/client-react/src/features/game/components/GameUI.tsx
@@ -1,14 +1,14 @@
 import { memo } from 'react'
-import { useApp } from '../../store/AppContext'
-import { useGameState, useCurrentPlayerMoves, useIsMyTurn, useIsConnected } from '../../store/gameSelectors'
-import GameBoard from '../game/GameBoard'
-import GameStatusBar from '../game/GameStatusBar'
-import GameLobby from '../game/GameLobby'
-import GameNavigation from '../game/GameNavigation'
-import DebugMenu from '../debug/DebugMenu'
-import { PerformanceDashboard } from '../debug/PerformanceDashboard'
-import { useGameLifecycle } from '../../hooks/useGameLifecycle'
-import { useDebugState } from '../../hooks/useDebugState'
+import { useApp } from '../../../store/AppContext'
+import { useGameState, useCurrentPlayerMoves, useIsMyTurn, useIsConnected } from '../store/gameSelectors'
+import GameBoard from './GameBoard'
+import GameStatusBar from './GameStatusBar'
+import GameLobby from '../../lobby/components/GameLobby'
+import GameNavigation from './GameNavigation'
+import DebugMenu from '../../debug/components/DebugMenu'
+import { PerformanceDashboard } from '../../debug/components/PerformanceDashboard'
+import { useGameLifecycle } from '../hooks/useGameLifecycle'
+import { useDebugState } from '../hooks/useDebugState'
 
 const GameUI = memo(() => {
   const gameState = useGameState()
