+++ packages/client-react/src/features/game/components/STLPiece.tsx
