+++ packages/client-react/src/features/game/components/board-types.ts
