+++ packages/client-react/src/features/game/hooks/useGameLifecycle.ts
