+++ packages/client-react/src/features/game/hooks/useReadyState.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from 'react'
 import { useGameState, useGameId } from '../store/gameSelectors'
-import { useApp } from '../store/AppContext'
+import { useApp } from '../../../store/AppContext'
 import { webSocketClient } from '../services/WebSocketClient'
 
 /**
