+++ packages/client-react/src/features/game/services/GameMessageHandler.ts
@@ -14,9 +14,9 @@
  */
 
 import { useGameStore } from '../store/gameStore'
-import { useAppStore } from '../store/AppContext'
+import { useAppStore } from '../../../store/AppContext'
 import type { AvailableMove, GameState, GameStateUpdatePayload, PlayerGameView } from '../types/game'
-import { WS_MESSAGE_TYPES } from '../../../shared/src/websocket-types'
+import { WS_MESSAGE_TYPES } from '../../../../../shared/src/websocket-types'
 
 // Type guard for PlayerGameView (has isCurrentPlayer field)
 function isPlayerGameView(view: GameStateUpdatePayload['state']): view is PlayerGameView {
