+++ packages/client-react/src/features/game/services/WebSocketService.ts
@@ -1,12 +1,12 @@
 import type { AvailableMove, GameState, GameStateUpdatePayload, PlayerGameView } from '../types/game'
-import { ErrorHandler, ErrorType } from '../utils/errorHandler'
+import { ErrorHandler, ErrorType } from '../../../utils/errorHandler'
 import { useGameStore } from '../store/gameStore'
-import { RECONNECTION_DELAY } from '../constants/gameConstants'
+import { RECONNECTION_DELAY } from '../../../constants/gameConstants'
 
 // Import shared WebSocket types
 import {
   WS_MESSAGE_TYPES
-} from '../../../shared/src/websocket-types'
+} from '../../../../../shared/src/websocket-types'
 
 interface WebSocketMessage {
   type: string;
@@ -52,7 +52,7 @@ export class WebSocketService {
 
     // Get JWT token from the server
     try {
-      const { apiService } = await import('./ApiService')
+      const { apiService } = await import('../../../services/ApiService')
       const tokenResponse = await apiService.getToken()
       this.authToken = tokenResponse.token
       console.log('🔑 Got JWT token for WebSocket:', this.authToken ? 'present' : 'missing')
