+++ packages/client-react/src/features/lobby/components/GameLobby.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react'
-import { useGameLifecycle } from '../../hooks/useGameLifecycle'
-import { useReadyState } from '../../hooks/useReadyState'
+import { useGameLifecycle } from '../../game/hooks/useGameLifecycle'
+import { useReadyState } from '../../game/hooks/useReadyState'
 import PlayerList from './PlayerList'
 
 interface GameLobbyProps {
